+++ libs/portal-integration-angular/src/index.ts
@@ -93,3 +93,8 @@ export * from './lib/core/primeng.module'
 
 // functions
 export * from './lib/functions/flatten-object'
+
+// utils
+export * from './lib/core/utils/objectutils'
+export * from './lib/core/utils/dateutils'
+export * from './lib/core/utils/translate.combined.loader'