+++ libs/portal-integration-angular/src/lib/model/column-type.model.ts
@@ -1,5 +1,6 @@
 export const enum ColumnType {
-  STANDARD = 'STANDARD',
+  STRING = 'STRING',
+  NUMBER = 'NUMBER',
   DATE = 'DATE',
   RELATIVE_DATE = 'RELATIVE_DATE',
   TRANSLATION_KEY = 'TRANSLATION_KEY',