+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+## [3.1.1](https://github.com/onecx/onecx-portal-ui-libs/compare/v3.1.0...v3.1.1) (2023-11-08)
+
+
+### Bug Fixes
+
+* add menu button tooltip ([044b7b7](https://github.com/onecx/onecx-portal-ui-libs/commit/044b7b7064d4cc7968a8ffd114b5d4bb9b1c4f36))
+* add missing imports to tests ([152eb96](https://github.com/onecx/onecx-portal-ui-libs/commit/152eb964f732c66602df957df640e13a52284a85))
+* export TranslateCombinedLoader ([5aeac59](https://github.com/onecx/onecx-portal-ui-libs/commit/5aeac59226e398af15a98e6fd3d77444b25bfaa5))
+
+## [3.1.0](https://github.com/onecx/onecx-portal-ui-libs/compare/v3.0.1...v3.1.0) (2023-11-08)
+
+
+### Features
+
+* added onDestroy to unsubscribe from services ([60ee997](https://github.com/onecx/onecx-portal-ui-libs/commit/60ee997655ae39312cf00210b4f245c34d342eb4))
+* added package-lock ([b7d0210](https://github.com/onecx/onecx-portal-ui-libs/commit/b7d0210bb60c69e1a1460708fb86fca007382589))
+* Added trigger for PRs ([e3523eb](https://github.com/onecx/onecx-portal-ui-libs/commit/e3523ebbc76bda54f42e3eca66cc6ad5c6a6e552))
+* removed !main ([96fc6cf](https://github.com/onecx/onecx-portal-ui-libs/commit/96fc6cf4f842f74826f703f59eff4ed2fc90793e))
+* trigger pipeline ([4dcbaa2](https://github.com/onecx/onecx-portal-ui-libs/commit/4dcbaa2906f8cbd81438b389f0613077e443d7f3))
+
+
+### Bug Fixes
+
+* add dep to top level ([6dfb07e](https://github.com/onecx/onecx-portal-ui-libs/commit/6dfb07eb42528edf1237bcd71c301dc87ac8ab4a))
+* correct dates in tests ([acc56e8](https://github.com/onecx/onecx-portal-ui-libs/commit/acc56e8622b4c601213418611df7def7b786d69b))
+* correct sorting of numbers in list grid ([50ab81d](https://github.com/onecx/onecx-portal-ui-libs/commit/50ab81d99d87d0e03259079747bae50217abfc95))
+* remove flattenItems in list-grid, export resolveFieldData ([6e4e010](https://github.com/onecx/onecx-portal-ui-libs/commit/6e4e01035340ef04317824909507a95a7e930bfe))
+
 ## [3.0.1](https://github.com/onecx/onecx-portal-ui-libs/compare/v3.0.0...v3.0.1) (2023-10-26)